Move to lib/datamodel from lib/core
