+++ src/lib/Makefile.am
@@ -37,6 +37,7 @@ include ../inet/InetLayer.am
 include ../system/SystemLayer.am
 include ../transport/TransportLayer.am
 include core/CoreLayer.am
+include datamodel/DataModel.am
 include support/SupportLayer.am
 include profiles/ProfilesLayer.am
 include ../crypto/Crypto.am
@@ -46,6 +47,9 @@ include ../app/DataModel.am
 CoreLayer_dir=$(includedir)/core
 dist_CoreLayer__HEADERS=$(CHIP_BUILD_CORE_LAYER_HEADER_FILES)
 
+# install headers for datamodel layer
+DataModel_dir=$(includedir)/datamodel
+dist_DataModel__HEADERS=$(CHIP_BUILD_DATAMODEL_LAYER_HEADER_FILES)
 # install headers for DeviceController
 Controller_dir=$(includedir)/controller
 dist_Controller__HEADERS=$(CHIP_BUILD_DEVICE_CONTROLLER_HEADER_FILES)
@@ -73,6 +77,7 @@ libCHIP_a_CPPFLAGS                  =                         \
 libCHIP_a_SOURCES                   = $(CHIP_BUILD_SYSTEM_LAYER_SOURCE_FILES)
 libCHIP_a_SOURCES                  += $(CHIP_BUILD_INET_LAYER_SOURCE_FILES)
 libCHIP_a_SOURCES                  += $(CHIP_BUILD_CORE_LAYER_SOURCE_FILES)
+libCHIP_a_SOURCES                  += $(CHIP_BUILD_DATAMODEL_LAYER_SOURCE_FILES)
 libCHIP_a_SOURCES                  += $(CHIP_BUILD_SUPPORT_LAYER_SOURCE_FILES)
 libCHIP_a_SOURCES                  += $(CHIP_BUILD_DEVICE_CONTROLLER_SOURCE_FILES)
 libCHIP_a_SOURCES                  += $(CHIP_BUILD_DATA_MODEL_SOURCE_FILES)
