+++ src/lib/datamodel/DataModel.am
@@ -1,0 +1,38 @@
+#
+#    Copyright (c) 2020 Project CHIP Authors
+#
+#    Licensed under the Apache License, Version 2.0 (the "License");
+#    you may not use this file except in compliance with the License.
+#    You may obtain a copy of the License at
+#
+#        http://www.apache.org/licenses/LICENSE-2.0
+#
+#    Unless required by applicable law or agreed to in writing, software
+#    distributed under the License is distributed on an "AS IS" BASIS,
+#    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#    See the License for the specific language governing permissions and
+#    limitations under the License.
+#
+
+#
+#    Description:
+#      This file is the GNU automake template for the CHIP 
+#      library DataModel sources.
+#
+#      These sources are shared by other SDK makefiles and consequently
+#      must be anchored relative to the top build directory.
+#
+
+CHIP_BUILD_DATAMODEL_LAYER_SOURCE_FILES                   = \
+    @top_builddir@/src/lib/datamodel/CHIPEndPoint.cpp       \
+    $(NULL)
+
+CHIP_BUILD_DATAMODEL_LAYER_HEADER_FILES                   = \
+    @top_builddir@/src/lib/datamodel/CHIPBaseAttribute.h         \
+    @top_builddir@/src/lib/datamodel/CHIPBaseCluster.h           \
+    @top_builddir@/src/lib/datamodel/CHIPClusterServer.h         \
+    @top_builddir@/src/lib/datamodel/CHIPEndPoint.h              \
+    @top_builddir@/src/lib/datamodel/CHIPStandardAttributes.h    \
+    @top_builddir@/src/lib/datamodel/CHIPStandardClusters.h      \
+    @top_builddir@/src/lib/datamodel/CHIPValues.h                \
+    $(NULL)
